python-ai-service/requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf==5.26.1
aiohttp==3.10.5
pydantic==2.7.1
python-dotenv==1.0.1
pydantic-settings==2.2.1
